// src/booking/database/nosql/model/array/deceased.array.ts
export class Deceased {

    id: string;
    deceasedPersonId: string;
    authoriserPersonId: string;
    salesId: string;
    siteId: string;
    casket: boolean;
    coffin: boolean;
    coffinType: string;
    detailsAndRequests: string;
    funeralCompanyId: string;
    functionOrCateringDateAndTime: Date;
    functionOrCateringRequested: string;
    funeralArranger: string;
    height: string;
    largeService: boolean;
    length: string;
    numberOfAttendees: string;
    overSized: boolean;
    payment: string;
    private: boolean;
    prePurchaseDetails: string;
    serviceDate: Date;
    serviceToBePaidBy: string;
    serviceType: string;
    sizeOfCoffinOrCasket: string;
    status: string;
    transactionType: string;
    weight: string;
    width: string;
    serviceVenueType: string;
    venueCompanyID: number;
    serviceCelebrant: number;

}